[REFACTOR]: dto/converter 분리 및 코드개선
