+++ pickme/src/main/java/project/pickme/notice/dto/NoticeDto.java
@@ -25,37 +25,4 @@ public class NoticeDto {
 	private String customsId;
 	private String customsName;
 
-	public NoticeDto withId(Long id) {
-		return NoticeDto.builder()
-			.id(id)
-			.title(title)
-			.content(content)
-			.createTime(createTime)
-			.customsId(customsId)
-			.customsName(customsName)
-			.build();
-	}
-
-	public Notice toEntity(Customs customs) {
-		return Notice.builder()
-			.id(id)
-			.title(title)
-			.content(content)
-			.createTime(createTime != null ? createTime : LocalDateTime.now())
-			.type(NoticeType.NOTICE)
-			.customs(customs)
-			.build();
-	}
-
-	public static NoticeDto fromEntity(Notice notice) {
-		Customs customs = notice.getCustoms();
-		return NoticeDto.builder()
-			.id(notice.getId())
-			.title(notice.getTitle())
-			.content(notice.getContent())
-			.createTime(notice.getCreateTime())
-			.customsId(customs.getId())
-			.customsName(customs.getName())
-			.build();
-	}
 }
